Fix audit LOW findings (overflow, guards, error msgs, validation)
Defensive hardening from the commit audit (none observed in practice; all LOW):

C core:
* tvdb_idx / tvdb_idx_d / tvdb_grid_voxels now use size_t arithmetic instead of
  int, so flat indices and voxel counts don't overflow for grids > INT_MAX
  voxels (the new comp_* ops went through the int helper). lvl_fill's row index
  likewise widened.
* grid_index: reject non-positive voxel_size in voxelize_points / points_in_set
  (return false) and guard world_to_ijk's divide (a zero size would make
  floorf(inf)->int cast UB).
* volume_render: fall back to an alternate up axis when `up` is parallel to the
  view direction (the cross product would collapse to a zero basis).

Python bindings:
* mod_sample_quadratic / mod_sample_trilinear: NULL-data check before use.
* tvdb_py_advect_scheme / tvdb_py_resample_grid: set s_error_msg on the
  malloc-failure path (was surfacing a stale message).
* spatial-query module fns (morton encode/decode, voxelize, coords/points_in_set,
  ijk_to_index, neighbor_counts): validate the input buffer length is a multiple
  of the element stride and raise ValueError instead of silently dropping
  trailing bytes.

19/19 ctests, 110 py passed.

+++ src/tinyvdb_ops.c
@@ -32,8 +32,8 @@ void tvdb_dense_vec_grid_free(tvdb_dense_vec_grid* grid) {
 // helpers
 // -------------------------------------------------------------------------
 
-static int tvdb_grid_voxels(const tvdb_dense_grid* g) {
-  return g->nx * g->ny * g->nz;
+static size_t tvdb_grid_voxels(const tvdb_dense_grid* g) {
+  return (size_t)g->nx * g->ny * g->nz;
 }
 
 static int tvdb_grid_same_shape(const tvdb_dense_grid* a, const tvdb_dense_grid* b) {
